Added contents for a second chapter, and it is very messy.
I'm trying to mix two old documents from past subjects into the thesis. One of them is review_text where I did a theoretical an state of the art review of LC and SLM characterization Methods.

The other document is that of the pre project. It has a part that was used for the introduction and also a part used for chapter 2.

The adition of these documents made the bibliography of each chapter much longer.

There is a lot of work to do at writting propper introductory texts and connectors between sections.